JS if-statements to check for var first, to save resources
Now menu.js, library.js, search.js all have if-statements to check for
a var before calling a method. This can help performance to keep the
browser from searching for a var through everything if it ends up not
existing… But if the var is going to be there, the if statement may
make it a tiny bit slower. So I keep the original “pre-if-check” files
as well, so that when I get to performance testing I can compare them.
NOTE: backarrow.js NEEDS if statements for performance, the way it is
set up now. Alternatively, the library.php, tag.php, and
blogarticle.php templates could each get an individualized version of
the backarrow.js script, containing on the portion of the original if
statement that pertains to them.