Back arrow hover text experimentation
Consider adding hover words to back arrow. The HTML and CSS files in
the GitHub “Back arrow hover text experimentation” save are all set up
for it, and there is one example for a simple version (just “Archive”)
in backarrow-post.js. The hold-up is collecting the page title of
referring links (when they are other blog posts). But here is an idea
for how to do it (and can be reworked to collect referring #tag pages,
library#entries, and (I think) search terms:
Use PHP and kirby cheat sheet methods to give each post title a
“data-all-post-titles” attribute that holds an array of every
blogarticle’s $page->title()->html() (or similar).
Use same method for another attribute: “data-all-post-URLs” to hold an
array of every blogarticle’s URL.
Or could use UID, and check via “include()” method if it is included in
the referredLink. This might not be as good for post titles, since I
could have a post that ends with “part 2” but is identical to the first
post (and the first post might not say “part 1”). But it is probably
better for Tag page and Search page.
In JS, find the URL attribute and split it (by whatever I used in PHP
to separate them with, e.g. a comma “,”).
Then loop through the resulting array, looking for one that matches the
referredLink.
Then collect that array item’s array count (e.g. [0]).
Split the post title attribute similarly.
Loop through the resulting post title array similarly.
Use the array count I got from the first array loop (through
“data-all-post-URLs” attribute) to find the corresponding count in the
second array loop (for “data-all-post-titles”) to get the
referredLink’s post title.
Resources for this solution:
GitHub “Back arrow hover text experimentation” July 1, 2017
⚯/JS/Back Arrow bookmarks