Vocab final pre-optimization
Everything works, but if the performance is too slow on click (to pull
up the inline definition, or close it, etc), I should consider
optimizing the script to pull all the content upon page load (and also
on resize if screen is 1225 and up). And change click behavior to just
show/hide the definition, instead of rebuilding/tearing it down on
every click.