Yellow menu can close without X button
Now you can tap anywhere on the screen where the yellow menu is NOT,
and it will close.

I added a <div> on all pages to surround ALL content except <nav> (and
except <scripts>), then added a JS event listener for when that <div>
was clicked, triggering it to close.