REAL Safari bug fix in-progress
Actually, Safari does not process any rems after it comes across
<details> tag…so I had to move the menu down to the last item in the
code, before the </body> closing tag, and use css to move its position
absolutely to the top (and the scripts that go with it). So far just
did this on the home blog.php page. Now need to do it for all.