p5 canvas can be opened in lightbox size at proportional size
And internal elements should be proportionally sized as well!
However, I had to resize the canvas upon lightbox open using regular JS. Not sure if this is the best approach due to:
// NOT CONVINCED THAT THIS IS THE BEST WAY TO FORCE CANVAS TO RESIZE
// CIRCLE ONLY RESIZES B/C blogarticle.js' draw() IS ALWAYS RE-CHECKING ANY OF ITS INTERNAL PARAMETERS
// AND I GAVE IT IF-STATEMENTS ABOUT SCREENWIDTH + CHECKING IF sketchHolder HAD CLASS OF .sketch-lightbox
// IT MIGHT BE MUCH BETTER TO HAVE blogarticle.js LISTEN FOR A CLICK ON FULLSCREENBUTTON,
// THEN TRIGGER A redraw(); AND/OR resizeCanvas(..., ...);

I couldn't figure out how to do this with p5 yet, but I am quite certain there is a way. I just need to learn the ins and outs of p5. When I figure out p5 better and discern the way to resize or re-draw, etc the canvas with it, I will return to this project, and will change it to use p5 only, if feasible/better. This is just a placeholder solution for now, for my exploration of if it was even possible to hack this together, according to Issue #176 .