<title> tag now reference's post page's "name" attribute, not sub-folder's "title" attribute
This was a relic bug after I fixed the search results by changing the page's "title" attribute (in its blueprint .yml file) to be called a "name" attribute instead. So I needed to point the <title> tag to the "name" attribute so it shows the post name in the browser tab (and when bookmarking).